+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/maps/MapController.java
@@ -8,16 +8,16 @@ import org.bukkit.inventory.ItemStack;
 
 public interface MapController {
     List<URLMap> getAll();
-    URLMap getMap(short id);
-    void loadMap(String world, short id, String url, String name, int x, int y, int width, int height, Integer priority);
+    URLMap getMap(int id);
+    void loadMap(String world, int id, String url, String name, int x, int y, int width, int height, Integer priority);
     ItemStack getURLItem(String world, String url, String name, int x, int y, int width, int height, Integer priority);
     void forceReloadPlayerPortrait(String worldName, String playerName);
     @Nullable
     ItemStack getPlayerPortrait(String worldName, String playerName, Integer priority, String photoName);
-    ItemStack getMapItem(short id);
-    boolean hasMap(short id);
-    boolean remove(short id);
+    ItemStack getMapItem(int id);
+    boolean hasMap(int id);
+    boolean remove(int id);
     void save();
-    short getURLMapId(String world, String url, String name, int x, int y, int width, int height, Integer priority);
-    short getURLMapId(String world, String url);
+    int getURLMapId(String world, String url, String name, int x, int y, int width, int height, Integer priority);
+    int getURLMapId(String world, String url);
 }
